Simplify over-engineered IcebergCommandArgumentsParser into direct parsing
Replace the generic IcebergCommandArgumentsParser framework (~400 lines,
7-variant ArgType enum, ParsedArguments class with ~15 getters) with:

- ExpireSnapshotsTypes.h: lightweight header for Options/Result structs,
  breaking the heavy Mutations.h transitive include from ExecuteOptionsParser.h
- IcebergFieldParseHelpers.h/cpp: four reusable free functions
  (fieldToInt64, fieldToBool, fieldToPeriodMs, fieldToInt64Array)
- Direct parseExpireSnapshotsOptions (~80 lines) that parses AST straight
  into ExpireSnapshotsOptions without an intermediate representation

Net: -665 lines, +45 lines.
Made-with: Cursor
