+++ src/Storages/ObjectStorage/DataLakes/Iceberg/Mutations.h
@@ -15,6 +15,7 @@
 #include <Parsers/ASTCreateQuery.h>
 #include <Parsers/ASTDropQuery.h>
 #include <Parsers/ASTLiteral.h>
+#include <Storages/ObjectStorage/DataLakes/Iceberg/ExpireSnapshotsTypes.h>
 #include <Storages/ObjectStorage/StorageObjectStorageSource.h>
 
 namespace DB::Iceberg
@@ -41,26 +42,6 @@ void alter(
     const DataLakeStorageSettings & data_lake_settings,
     PersistentTableComponents & persistent_table_components,
     const String & write_format);
-
-struct ExpireSnapshotsResult
-{
-    Int64 deleted_data_files_count = 0;
-    Int64 deleted_position_delete_files_count = 0;
-    Int64 deleted_equality_delete_files_count = 0;
-    Int64 deleted_manifest_files_count = 0;
-    Int64 deleted_manifest_lists_count = 0;
-    Int64 deleted_statistics_files_count = 0;
-    bool dry_run = false;
-};
-
-struct ExpireSnapshotsOptions
-{
-    std::optional<Int64> expire_before_ms;
-    std::optional<Int64> retention_period_ms;
-    std::optional<Int32> retain_last;
-    std::optional<std::vector<Int64>> snapshot_ids;
-    bool dry_run = false;
-};
 
 ExpireSnapshotsResult expireSnapshots(
     const ExpireSnapshotsOptions & options,
